requirements.txt
Django==5.0.1
djangorestframework==3.14.0
pytest==8.0.0
pytest-django==4.8.0